refactor: split Phase 4 STM32H7 platform .h files into .c + .h pairs
Move implementations from clock.h, peripherals.h, lluart.h, llspi.h,
llfan.h, llflash.h, lladc.h, llfdcan.h, llusb.h, and sound.h into
corresponding .c files.
